Switch predicate initialization at extended level
