+++ src/main/java/seedu/address/logic/Messages.java
@@ -1,4 +1,3 @@
-
 package seedu.address.logic;
 
 import java.util.Set;
@@ -21,6 +20,7 @@ public class Messages {
     public static final String MESSAGE_VENDOR_NOT_ASSIGNED_TO_EVENT = "Vendor is not assigned to this event";
     public static final String MESSAGE_VENDOR_ALREADY_ASSIGNED = "Vendor is already assigned to this event";
     public static final String MESSAGE_VENDORS_LISTED_OVERVIEW = "%1$d vendors listed!";
+    public static final String MESSAGE_EVENTS_LISTED_OVERVIEW = "%1$d events listed!";
     public static final String MESSAGE_DUPLICATE_FIELDS =
         "Multiple values specified for the following single-valued field(s): ";
